Convert test output logging test to new JUnit multiversion test
